+++ CHANGES.md
@@ -7,6 +7,7 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
 
 ## next (not yet released)
 
+- Added support for the new `accessToken` option in m3api.
 - Updated dependencies.
 
 ## v0.3.3 (2025-03-27)